backend/requirements.txt
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
qdrant-client>=1.7.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
langchain-community>=0.0.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.0
python-multipart>=0.0.6
passlib>=1.7.0
python-jose[cryptography]>=3.3.0